+++ packages/site-demo/content/product/components/slideshow/react/default.tsx
@@ -40,8 +40,10 @@ export const App = ({ theme }: any) => {
     return (
         <Slideshow
             activeIndex={0}
-            nextButtonProps={{ label: 'Next' }}
-            previousButtonProps={{ label: 'Previous' }}
+            slideshowControlsProps={{
+                nextButtonProps: { label: 'Next' },
+                previousButtonProps: { label: 'Previous' },
+            }}
             theme={theme}
             groupBy={1}
             style={{ width: '50%' }}