+++ packages/site-demo/content/product/components/slideshow/react/group.tsx
@@ -19,8 +19,10 @@ export const App = ({ theme }: any) => {
     return (
         <Slideshow
             activeIndex={0}
-            nextButtonProps={{ label: 'Next' }}
-            previousButtonProps={{ label: 'Previous' }}
+            slideshowControlsProps={{
+                nextButtonProps: { label: 'Next' },
+                previousButtonProps: { label: 'Previous' },
+            }}
             theme={theme}
             autoPlay
             groupBy={2}